api/requirements.txt
# Dépendances Python pour l'API
flask==3.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
